Removed ban petitions
One mistake slipped through the cracks as I tested this:

It is not possible in any way for a user to DM a bot if they do not
share a guild. As of this moment, there is no workaround, and it
appears that simply opening a DM channel beforehand is not enough to
make it work.
In short: it doesn't work. It can't possibly work.

This feature has been removed. Alterations to the configuration
structure for ModTools will remain, as I planned to do this change
for other reasons not related to this feature anyway.